Food ingredient mvc design complete with aggregate admin api

+++ backend/src/main/java/com/cyro/craveKart/controllers/RestaurantAdminController.java
@@ -1,20 +1,18 @@
 package com.cyro.cravekart.controllers;
 
-
+import com.cyro.cravekart.config.security.AuthService;
 import com.cyro.cravekart.models.Restaurant;
 import com.cyro.cravekart.models.User;
+import com.cyro.cravekart.repository.RestaurantRepository;
+import com.cyro.cravekart.repository.UserRepository;
 import com.cyro.cravekart.request.CreateRestaurantRequest;
 import com.cyro.cravekart.response.CreateRestaurantResponse;
 import com.cyro.cravekart.service.RestaurantService;
-import com.cyro.cravekart.service.RestaurantServiceImpl;
 import com.cyro.cravekart.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -24,13 +22,15 @@ import java.util.List;
 @RequiredArgsConstructor
 @Slf4j
 public class RestaurantAdminController {
-
-  private final RestaurantServiceImpl restaurantService;
+  private final RestaurantRepository restaurantRepository;
+  private final RestaurantService restaurantService;
   private final UserService userService;
+  private final AuthService authService;
+  private final UserRepository userRepository;
 
-  @GetMapping("/health")
-  public ResponseEntity<?> getHealth() {
-    return new ResponseEntity<>(HttpStatus.OK);
+  @GetMapping("/check")
+  public ResponseEntity<String> getHealth() {
+    return new ResponseEntity<>( "OK", HttpStatus.OK);
   }
 
   @PostMapping
@@ -38,16 +38,12 @@ public class RestaurantAdminController {
       @RequestBody CreateRestaurantRequest request,
       @RequestHeader("Authorization") String jwtToken) {
 
-    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-    if (authentication == null ||
-        !authentication.isAuthenticated() ||
-        authentication.getPrincipal() instanceof String) {
-      return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+    User user = authService.getCurrentAuthUser();
+    if (user == null) {
+      return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
-    User user = (User) authentication.getPrincipal();
-    log.info("username while creating restaurant", user.getUsername());
+    log.info("username while creating restaurant :{}", user.getUsername());
     com.cyro.cravekart.response.CreateRestaurantResponse restaurantResponse = restaurantService.createRestaurant(request, user);
 
     return new ResponseEntity<>(restaurantResponse
@@ -59,10 +55,7 @@ public class RestaurantAdminController {
       @PathVariable Long restaurantId,
       @RequestBody CreateRestaurantRequest request){
 
-    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-    log.info("user trying to update restaurant " + restaurantId);
-    if(authentication != null || authentication.isAuthenticated() ||
-    authentication.getPrincipal() instanceof User) {
+    if(authService.getCurrentAuthUser() != null) {
 
       Restaurant updatedRestaurant = restaurantService.updateRestaurant(restaurantId, request);
       return  new ResponseEntity<>(updatedRestaurant, HttpStatus.OK);
@@ -71,37 +64,21 @@ public class RestaurantAdminController {
   }
 
   @DeleteMapping("/{restaurantId}")
-  public ResponseEntity<String> deleteRestaurant(
-      @PathVariable Long restaurantId
-  ){
-    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-    if(authentication != null || authentication.isAuthenticated() ||
-        authentication.getPrincipal() instanceof User) {
-      User user = (User) authentication.getPrincipal();
-      List<Restaurant> restaurantsByUserId = restaurantService.getRestaurantsByUserId(user.getId());
-      for(Restaurant restaurant : restaurantsByUserId) {
-        if(restaurant!= null) {
+  public ResponseEntity<String> deleteRestaurant(@PathVariable Long restaurantId){
+
+    User user = authService.getCurrentAuthUser();
+    if( user == null )
+      return  new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
+    List<Restaurant> restaurantsByUserId =
+        restaurantService.getRestaurantsByUserId(user.getId());
+
+    for(Restaurant restaurant : restaurantsByUserId) {
+        if(restaurant!= null && restaurant.getId().equals(restaurantId)) {
           restaurantService.deleteRestaurant(restaurant.getId());
           return new ResponseEntity<>("Deleted Successfully", HttpStatus.OK);
         }
       }
+    return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
-    return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
   }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-}
